Comments on the AsyncStateMachine interface.
