+++ src/partest-extras/scala/tools/partest/async/Async.scala
@@ -30,17 +30,6 @@ object Async {
   }
 }
 
-// The async phase expects the state machine class to structurally conform to this interface.
-trait AsyncStateMachine[F, R] {
-  protected def state$async_=(i: Int): Unit
-  protected def state$async: Int
-  protected def completeFailure(t: Throwable): Unit
-  protected def completeSuccess(value: AnyRef): Unit
-  protected def onComplete(f: F): Unit
-  protected def getCompleted(f: F): R
-  protected def tryGet(tr: R): AnyRef
-}
-
 abstract class AsyncAsMacroStateMachine(execContext: ExecutionContext) extends AsyncStateMachine[Future[AnyRef], Try[AnyRef]] with Function1[Try[AnyRef], Unit] {
   Objects.requireNonNull(execContext)
